# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/categories": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "カテゴリ一覧取得",
        "operationId": "get_categories_api_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCategoriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リソースが見つかりません。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "categories"
        ],
        "summary": "カテゴリ作成",
        "operationId": "create_category_api_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リソースが見つかりません。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      }
    },
    "/api/categories/{category_id}": {
      "put": {
        "tags": [
          "categories"
        ],
        "summary": "カテゴリ更新",
        "operationId": "update_category_api_categories__category_id__put",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCategoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リソースが見つかりません。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "categories"
        ],
        "summary": "カテゴリ削除",
        "operationId": "delete_category_api_categories__category_id__delete",
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Category Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リソースが見つかりません。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dashboards/summary": {
      "get": {
        "tags": [
          "dashboards"
        ],
        "summary": "ダッシュボード集計取得",
        "operationId": "get_dashboard_summary_api_dashboards_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryResponse"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboards/trends": {
      "get": {
        "tags": [
          "dashboards"
        ],
        "summary": "見積もり推移（月別）取得",
        "operationId": "get_dashboard_trends_api_dashboards_trends_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "開始日 (未指定の場合は6ヶ月前)",
              "title": "Start Date"
            },
            "description": "開始日 (未指定の場合は6ヶ月前)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "終了日 (未指定の場合は今日)",
              "title": "End Date"
            },
            "description": "終了日 (未指定の場合は今日)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardGraphResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエスト内容が不正です。"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "example": {
                  "detail": "リクエストを処理できません。"
                }
              }
            }
          },
          "500": {
            "description": "System Error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "システム管理者にお問い合わせください。"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateCategoryRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateCategoryRequest"
      },
      "CreateCategoryResponse": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ItemCategoryRead"
          }
        },
        "type": "object",
        "required": [
          "category"
        ],
        "title": "CreateCategoryResponse"
      },
...
      "DashboardGraphResponse": {
        "properties": {
          "graph_items": {
            "items": {
              "$ref": "#/components/schemas/GraphItem"
            },
            "type": "array",
            "title": "Graph Items"
          }
        },
        "type": "object",
        "required": [
          "graph_items"
        ],
        "title": "DashboardGraphResponse"
      },
      "DashboardSummaryResponse": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "current_month_count": {
            "type": "integer",
            "title": "Current Month Count"
          },
          "total_amount": {
            "type": "integer",
            "title": "Total Amount"
          },
          "average_amount": {
            "type": "integer",
            "title": "Average Amount"
          }
        },
        "type": "object",
        "required": [
          "total_count",
          "current_month_count",
          "total_amount",
          "average_amount"
        ],
        "title": "DashboardSummaryResponse"
      },
...
      "GetCategoriesResponse": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/ItemCategoryRead"
            },
            "type": "array",
            "title": "Categories"
          }
        },
        "type": "object",
        "required": [
          "categories"
        ],
        "title": "GetCategoriesResponse"
      },
...
      "GraphItem": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "month",
          "count",
          "amount"
        ],
        "title": "GraphItem"
      },
      "ItemCategoryRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "グループ名"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "is_deleted": {
            "type": "boolean",
            "title": "Is Deleted"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "is_deleted"
        ],
        "title": "ItemCategoryRead"
      },
...
      "UpdateCategoryRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "UpdateCategoryRequest"
      },
      "UpdateCategoryResponse": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ItemCategoryRead"
          }
        },
        "type": "object",
        "required": [
          "category"
        ],
        "title": "UpdateCategoryResponse"
      },
...
    }
  }
}